Extract plugin until-build verification into a separate class
